+++ ui/packages/shared/profile/src/components/ActionButtons/GroupByDropdown.tsx
@@ -180,7 +180,7 @@ const GroupByDropdown: React.FC<GroupByDropdownProps> = ({
 
   return (
     <div className="relative">
-      <label className="text-sm">Group</label>
+      <label className="text-sm">Group by</label>
       <div className="relative text-left" id="h-group-by-filter">
         <Button
           variant="neutral"